Removed separate suite for opensearch dashboard in post-upgrade
